requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiohttp-retry==2.9.1
aiosignal==1.4.0
asgiref==3.7.2
attrs==25.3.0
certifi==2023.11.17
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==41.0.7
defusedxml==0.7.1
Django==4.2.4
django-auth-style==0.8.1
django-cors-headers==4.3.1
django-jazzmin==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.0
frozenlist==1.7.0
gunicorn==21.2.0
idna==3.6
multidict==6.6.3
packaging==23.2
Pillow==10.0.1
propcache==0.3.2
psycopg2-binary==2.9.10
pycparser==2.21
PyJWT==2.8.0
PyMuPDF==1.26.3
python-decouple==3.8
pytz==2023.3.post1
razorpay==1.4.2
requests==2.31.0
sqlparse==0.4.4
stripe==7.7.0
twilio==9.6.4
typing_extensions==4.8.0
tzdata==2025.2
urllib3==2.1.0
whitenoise==6.9.0
yarl==1.20.1